requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
black==25.1.0
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
deepl==1.21.0
distro==1.9.0
executing==2.2.0
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
isort==6.0.1
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
narwhals==1.30.0
nest-asyncio==1.6.0
numpy==2.2.3
openai==1.63.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybars3==0.9.7
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMeta3==0.5.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
